fixed compilation warnings on Linux

+++ src/proto_holding_registers.cpp
@@ -47,7 +47,7 @@ rclcpp::FutureReturnCode Implementation::holding_register_read_handler_real_(
     case fc:
       // See if we have amount of data that is consistent with length
       response->len = result[1];
-      if ((response->len & 1) || result.length() != 2 + response->len) {
+      if ((response->len & 1) || result.length() != 2 + (size_t)response->len) {
         return rclcpp::FutureReturnCode::INTERRUPTED;
       }
